// frontend/src/generated/types/storeInfoRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */
import type { LocalTime } from './localTime';

export interface StoreInfoRequest {
  sysId?: string;
  storeName?: string;
  address?: string;
  phoneNumber?: number;
  emergencyPhoneNumber?: number;
  email?: string;
  openTime?: LocalTime;
  closeTime?: LocalTime;
}